noahdarveau/dialog treeshaking changes (#2576)
* Updated `dialog` capability to be treeshakable
