+++ change/@microsoft-teams-js-35af1950-5f26-4ffe-8c68-6ca569870710.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Made the `dialog` file treeshakable",
+  "packageName": "@microsoft/teams-js",
+  "email": "109628470+noahdarveau-MSFT@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
